fix: refactor structure and redesign add modal
