+++ src/components/Admin/TableManagement/index.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Table } from 'antd'
+
+function TableManagement({ rowSelection, columns, data }) {
+  return (
+    <Table
+      rowSelection={rowSelection}
+      columns={columns}
+      dataSource={data}
+      expandable={{
+        expandedRowRender: (record) => 'abc',
+      }}
+    />
+  )
+}
+
+export default TableManagement
