+++ src/components/index.js
@@ -1,6 +1,11 @@
+//Admin
 export { default as Sidebar } from './Admin/Sidebar'
 export { default as HeaderComponent } from './Admin/HeaderComponent'
-export { default as Bicycle } from './Admin/Bicycle'
+export { default as TableManagement } from './Admin/TableManagement'
+//BicycleManagement
+export { default as BicycleManageAction } from './Admin/BicycleManageAction'
+export { default as AddBicycleModal } from './Admin/AddBicycleModal'
+
 export { default as ImgCarousel } from './ImgCarousel'
 export { default as HomeNavigation } from './HomeNavigation'
 export { default as Logo } from './Logo'
